+++ memeCoinWebApp/Models/Transfer.cs
@@ -9,7 +9,7 @@ public class Transfer
     public int Id { set; get; }
     public string Source { set; get; }
     public string Destination { set; get; }
-    public int Amount { set; get; }
+    public decimal Amount { set; get; }
     public DateTime Timestamp { set; get; }
 
     // foreign keys
